+++ spec2nii/spec2nii.py
@@ -550,16 +550,8 @@ class spec2nii:
     # Bruker 2dseq files with FG_COMPLEX
     def bruker(self, args):
         from spec2nii.bruker import read_bruker
-        from warnings import warn
-        from brukerapi.exceptions import MissingProperty
-        warn('Bruker conversion must currently be run with numpy<1.20.0')
 
-        try:
-            self.imageOut, self.fileoutNames = read_bruker(args)
-        except MissingProperty:
-            raise Spec2niiError(
-                'Bruker conversion must currently be run with numpy<1.20.0. '
-                'The underlying brukerapi package requires updating for ongoing compatibility.')
+        self.imageOut, self.fileoutNames = read_bruker(args)
 
     # Varian parser
     def varian(self, args):